refactor: modify path name
