+++ src/app/(home)/_ui/loading.module.scss
