+++ src/app/(home)/template.tsx
@@ -5,8 +5,8 @@ import { ReactNode, Fragment } from "react";
 import { useAtomValue } from "jotai";
 import { loadingAtom } from "@/state/atom/loading";
 
-import Background from "./_components/background";
-import Loading from "./_components/loading";
+import Background from "./_ui/background";
+import Loading from "./_ui/loading";
 
 export default function HomeTemplate({ children }: { children: ReactNode }) {
   const { isStarted } = useAtomValue(loadingAtom);
